# swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
      "version": "1.0.0",
      "title": "Train Location Service API",
      "description": "API for tracking train locations and querying the latest location."
    },
    "host": "localhost:3001",
    "basePath": "/",
    "schemes": [
       "http",
       "https"
],
    "consumes": ["application/json"],
    "produces": ["application/json"],
    "paths": {
      "/update-location": {
        "post": {
          "summary": "Update Train Location",
          "description": "Updates the location of a train with latitude and longitude.",
          "parameters": [
            {
              "name": "body",
              "in": "body",
              "required": true,
              "description": "Train location data",
              "schema": {
                "type": "object",
                "properties": {
                  "trainId": {
                    "type": "string",
                    "example": "12345",
                    "description": "Unique ID of the train"
                  },
                  "latitude": {
                    "type": "number",
                    "example": 6.9271,
                    "description": "Latitude of the train"
                  },
                  "longitude": {
                    "type": "number",
                    "example": 79.8612,
                    "description": "Longitude of the train"
                  }
                },
                "required": ["trainId", "latitude", "longitude"]
              }
            }
          ],
          "responses": {
            "201": {
              "description": "Location data saved successfully"
            },
            "400": {
              "description": "Bad Request"
            },
            "500": {
              "description": "Error saving location data"
            }
          }
        }
      },
      "/train-locations/{trainId}/latest": {
        "get": {
          "summary": "Get Latest Train Location",
          "description": "Retrieves the latest location of a train based on its ID.",
          "parameters": [
            {
              "name": "trainId",
              "in": "path",
              "required": true,
              "type": "string",
              "description": "Unique ID of the train"
            }
          ],
          "responses": {
            "200": {
              "description": "Latest location of the train"
            },
            "404": {
              "description": "Train location not found"
            },
            "500": {
              "description": "Error retrieving location data"
            }
          }
        }
      }
    }
  }